Fix: Handle single EIC case for merged EIC #334
For a single sample, merged EIC and spline should coincide. grouPeaks was handling the case for a single sample by not calling eicMerge and creating its spline.

addMergedEIC on the other hand, did not handle the single sample case and therefore a merged EIC was created and smoothed leading to different spline and mergedEIC.

Issue: #334